+++ spec/lite/uxid/ulid_spec.rb
@@ -17,28 +17,4 @@ RSpec.describe Lite::Uxid::Ulid do
     end
   end
 
-  describe '.uxid_bytes' do
-    it 'to be 16' do
-      expect(encoder.uxid_bytes.length).to eq(16)
-    end
-  end
-
-  describe '.uxid_octect' do
-    it 'to be a random string' do
-      expect(encoder.uxid_octect).not_to eq(nil)
-    end
-  end
-
-  describe '.uxid_unixtime_flex' do
-    it 'to be a random integer' do
-      expect(encoder.uxid_unixtime_flex).not_to eq(nil)
-    end
-  end
-
-  describe '.uxid_unixtime_48bit' do
-    it 'to be 6' do
-      expect(encoder.uxid_unixtime_48bit.length).to eq(6)
-    end
-  end
-
 end